Solve problem 28
There is a more mathematical solution - by recognizing that the upper
right corner is n² and each corner is just n-1 from the previous. That
yields a layer sum of 4n² - 6n + 6. That equation could be calculated
for each layer and summe, or for ultra speed, partial sums could be used
(something like 4∑n² - 6∑n-1).