Improve performance of problem 39
This solution is largely based on rayfil's explanation, which simply
directly determines the only value of b for any given value of a.

While not "my own work", it was helpful to me to walk through the
mathematical proof of his equation to find b when given a perimeter and
the length of one side.

This produces the exact same sum of solutions as my original solution as
further proof that if b is an integer, c would be as well. This makes
sense because c in this, is just P-a-b.